requirements.txt
fastapi
uvicorn
python-multipart
python-jose
passlib
streamlit==1.20.0
# sentence-transformers==2.2.2
pinecone-client==2.2.1
uuid
openai==0.27.2
bcrypt
mangum